Add top-level and formspec-level `init.lua` files
Top-level `init.lua` exposes libuix as a valid mod and adds a
`libuix` class that returns a per-mod instance of the library and
all its tools.

Formspec-level `init.lua` collects the important modules located in
the `formspec` directory into a single table and exposes it to the
top-level of the project for the sake of simplicity.