Optimize elements list generation via Builder class
Now when a variation is registered, it enters the elements list as an
instance of the Variation class rather than an instance of the Element
class with only one Variation. When a second variation of the same
element is registered the Variation entry is converted to an Element.
This eliminates some additional runtime logic introduced by the Element
class and prevents the need to create and store instances of both the
Element and Variation classes