Add error blame level and use ErrorBuilder throughout
Error blame level simplifies the traceback by blaming functions further
up the call stack. It can't be used in every case but for the many that
it can it makes reading errors much simpler.

ErrorBuilder is now used throughout the program in place of direct calls
to `error`. It may be worth expanding unit testing to force more errors
for the Variation, Element, and Form classes.

Every instance of ErrorBuilder is now created outside of the function
using it to increase execution speed by preventing the unnecessary
repeated creation of instances.